Move adding default chart to a separate saga to make sure it triggers layout update.
